requirements.txt
selenium==4.23.1
webdriver-manager==4.0.2
pytest==7.4.4
pytest-html==3.1.1
allure-pytest==2.13.2
pandas==2.2.2
openpyxl==3.1.5
pyyaml==6.0.2
pymysql==1.1.1